parte adm 50%

+++ src/main/java/com/quiz/quiz_sistema_solar/repository/PerguntaRepository.java
@@ -1,9 +1,10 @@
 package com.quiz.quiz_sistema_solar.repository;
 
-import com.quiz.quiz_sistema_solar.model.Pergunta;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.quiz.quiz_sistema_solar.model.Pergunta;
+
 @Repository
-public interface PerguntaRepository extends JpaRepository<Pergunta, Long> {
+public interface PerguntaRepository extends JpaRepository<Pergunta, Integer> {
 }
